third_party/blink/web_tests/platform/generic/fast/writing-mode/auto-margins-across-boundaries-expected.txt

The green LR block should be centered horizontally within the black-bordered block.
PASS: The block is in the correct position.
